cleanup duplicate code, add comments
1. Created autoSizeUtils.ts - A new shared utility file containing:

    - AutoSizeContext interface (now imported from types.ts)
    - ensureContext() - Ensures the context object exists
    - ensureAutoSizeContext() - Ensures the autoSize context exists
    - currentAutoSizeContext() - Gets the current autoSize context or null
    - cleanupAutoSizeContext() - Cleans up empty autoSize context

2. Updated scaleLayoutItemSizes.ts:

    - Removed duplicate AutoSizeContext interface
    - Removed duplicate helper functions (ensureContext, ensureAutoSizeContext, currentAutoSize, cleanupAutoSizeContext)
    - Now imports and uses helpers from autoSizeUtils.ts
    - Fixed function call from currentAutoSize to currentAutoSizeContext for consistency

3. Updated setLayoutItemProps.ts:

    - Removed duplicate AutoSizeContext interface
    - Removed duplicate helper functions (ensureContext, ensureAutoSize, cleanupAutoSize)
    - Now imports and uses helpers from autoSizeUtils.ts

4. Updated types.ts:

    - Added AutoSizeContext interface (single source of truth)
    - Added LayoutItemContext interface
    - Properly typed LayoutItem.context property
