+++ .eslintignore
@@ -396,6 +396,7 @@ packages/app-desktop/gui/PromptDialog.js
 packages/app-desktop/gui/ResizableLayout/LayoutItemContainer.js
 packages/app-desktop/gui/ResizableLayout/MoveButtons.js
 packages/app-desktop/gui/ResizableLayout/ResizableLayout.js
+packages/app-desktop/gui/ResizableLayout/utils/autoSizeUtils.js
 packages/app-desktop/gui/ResizableLayout/utils/findItemByKey.js
 packages/app-desktop/gui/ResizableLayout/utils/isTempContainer.js
 packages/app-desktop/gui/ResizableLayout/utils/iterateItems.js
