+++ packages/app-desktop/gui/ResizableLayout/utils/types.ts
@@ -8,6 +8,20 @@ export interface Size {
 	height: number;
 }
 
+export interface AutoSizeContext {
+	naturalWidth?: number;
+	naturalHeight?: number;
+	rootWidth?: number;
+	rootHeight?: number;
+}
+
+export interface LayoutItemContext {
+	savedRootSize?: Size;
+	autoSize?: AutoSizeContext;
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Old code before rule was applied
+	[key: string]: any;
+}
+
 export interface LayoutItem {
 	key: string;
 	isRoot?: boolean;
@@ -20,8 +34,7 @@ export interface LayoutItem {
 	resizableRight?: boolean;
 	resizableBottom?: boolean;
 	visible?: boolean;
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Old code before rule was applied
-	context?: any;
+	context?: LayoutItemContext;
 }
 
 export const tempContainerPrefix = 'tempContainer-';
